move to svm_message_ext
